requirements.txt
astroid==1.6.3
certifi==2018.4.16
chardet==3.0.4
codecov==2.0.15
coverage==4.0.3
flake8==3.5.0
idna==2.6
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.2.0
pycodestyle==2.3.1
pyflakes==1.6.0
pylint==1.8.4
python-interface==1.4.0
pytz==2018.5
requests==2.18.4
rope==0.11.0
six==1.11.0
urllib3==1.22
wrapt==1.10.11
